subs/toasty

+++ src/app/borrowings/borrowings.component.ts
@@ -4,7 +4,13 @@ import {Borrowing} from "../model/borrowing.model";
 import {UserService} from "../service/user/user.service";
 import {User} from "../model/user.model";
 import {BorrowingService} from "../service/borrowing/borrowing.service";
+import {Book} from "../model/book.model";
+import {BookService} from "../service/book/book.service";
+import {UntilDestroy, untilDestroyed} from "@ngneat/until-destroy";
+import {Subscription} from "rxjs";
+import {ToastService} from "angular-toastify";
 
+@UntilDestroy()
 @Component({
   selector: 'app-borrowings',
   templateUrl: './borrowings.component.html',
@@ -16,10 +22,18 @@ export class BorrowingsComponent {
   borrowing?: Borrowing;
 
   users?: User[];
+  books?: Book[];
 
-  constructor(private borService: BorrowingService, private usService: UserService) {
+  private getListSubscription?: Subscription;
+
+  constructor(private borService: BorrowingService, private usService: UserService, private bookService: BookService, private toastService: ToastService) {
     this.borService.getBorrowings();
     this.getUsers();
+    this.getBooks();
+  }
+
+  ngOnDestroy(): void {
+    this.getListUnsubscribe();
   }
 
   createBorrowing(borrowing: Borrowing) {
@@ -43,14 +57,18 @@ export class BorrowingsComponent {
   }
 
   deleteBorrowing(borrowingId: number): void {
-    this.borService.deleteBorrowing(borrowingId).subscribe(() => {
-      console.log("Delete borrowing OK");
-      this.getBorrowings();
-    })
+    if(window.confirm("Naozaj chcete vymazat vypozicku?")){
+      this.borService.deleteBorrowing(borrowingId).subscribe(() => {
+        console.log("Delete borrowing OK");
+        this.getBorrowings();
+      }, () => {
+        this.toastService.error("Chyba. Vypozicka nebola vymazana.");
+      })
+    }
   }
 
   private getBorrowings() {
-    this.borService.getBorrowings().subscribe((borrowings: Borrowing[]) => {
+    this.borService.getBorrowings().pipe(untilDestroyed(this)).subscribe((borrowings: Borrowing[]) => {
       this.borrowings = borrowings;
     })
   }
@@ -59,5 +77,18 @@ export class BorrowingsComponent {
     this.usService.getUsers().subscribe(users => {
       this.users = users;
     })
+  }
+
+  private getBooks(): void {
+    this.bookService.getBooks().subscribe(books => {
+      this.books = books;
+    })
+  }
+
+  private getListUnsubscribe() {
+    if(this.getListSubscription) {
+      this.getListSubscription.unsubscribe();
+      this.getListSubscription = undefined;
+    }
   }
 }
